map_handling: Moved Direction to map_util
