+++ python/day6/part1.py
@@ -1,6 +1,6 @@
-from day6.utils import init_maze_findings, Direction
+from day6.utils import init_maze_findings, turn_right
 from file_handling import get_file_as_lines
-from map_handling import Coordinate
+from map_handling import Coordinate, Direction
 
 
 def main(lines: list[str]) -> int:
@@ -38,7 +38,7 @@ def get_next_guard_state_with_obstacles(guard_position: Coordinate, guard_direct
     next_position = calc_next_position(guard_direction, guard_position)
     if next_position not in obstacle_positions:
         return next_position, guard_direction
-    return guard_position, guard_direction.turn_right()
+    return guard_position, turn_right(guard_direction)
 
 
 def calc_next_position(guard_direction: Direction, guard_position: Coordinate) -> Coordinate:
